requirements.txt
Django==1.11.1
channel
requests
